Refactored easyMode and normalMode methods into helper files and renamed them to be more descriptie
