+++ src/NormalMode.js
@@ -1,27 +1,26 @@
 import React, { useState, useEffect } from "react";
-import { DIAGONALS } from "./data/diagonals";
+import { THREE_BY_THREES } from "./data/threeByThree";
 import HowToPlay from "./components/HowToPlay";
 import Grid from "./components/Grid";
 import Score from "./components/Score";
 import LetterBucket from "./components/LetterBucket";
+import { parseWords, getDayNumber } from "./helpers/Helpers";
 import {
-  generateRowNeighbours,
-  generateEasyGrid,
-  parseWords,
-  getDayNumber,
-} from "./Helpers";
+  generateRowColNeighbours,
+  generateNormalGrid,
+} from "./helpers/normalModeMethods";
 
-console.log(DIAGONALS[getDayNumber() - 1]);
+console.log(THREE_BY_THREES[getDayNumber() - 1]);
 
-export default function DiagonalMode() {
+export default function NormalMode() {
   const MAXSCORE = 25;
   const WORDLENGTH = 5;
-  const rawDailyAnswer = DIAGONALS[getDayNumber() - 1];
+  const rawDailyAnswer = THREE_BY_THREES[getDayNumber() - 1];
   const parsedDailyAnswer = parseWords(rawDailyAnswer);
 
-  const initialGrid = generateEasyGrid(parsedDailyAnswer);
+  const initialGrid = generateNormalGrid(parsedDailyAnswer);
   const possibleLetters = parsedDailyAnswer;
-  let eachCellsNeighbours = generateRowNeighbours(initialGrid);
+  let eachCellsNeighbours = generateRowColNeighbours(initialGrid);
 
   const [grid, setGrid] = useState(initialGrid);
   const [showModal, setShowModal] = useState(false);
@@ -42,7 +41,7 @@ export default function DiagonalMode() {
   const changeCell = (e) => {
     const input = e.target.value.toUpperCase();
     if (!/^[a-zA-Z]*$/.test(input)) {
-      return false;
+      e.target.preventDefault();
     }
     const r = e.target.attributes.row.value;
     const c = e.target.attributes.col.value;
@@ -63,11 +62,12 @@ export default function DiagonalMode() {
           eachCellsNeighbours.rows[i].cols[j].neighbours.includes(cell.value)
         ) {
           newGrid.rows[i].cols[j].state = "wrong-location";
+        } else if (cell.answer === "*") {
         } else {
           newGrid.rows[i].cols[j].state = "wrong";
         }
         updateCorrectCellCount(cell);
-        eachCellsNeighbours = generateRowNeighbours(newGrid);
+        eachCellsNeighbours = generateRowColNeighbours(newGrid);
       }
     }
     setGrid(newGrid);
