+++ src/helpers/normalModeMethods.js
@@ -1,51 +1,3 @@
-function getDayNumber() {
-  let now = new Date();
-  let start = new Date(now.getFullYear(), 0, 0);
-  let diff =
-    now -
-    start +
-    (start.getTimezoneOffset() - now.getTimezoneOffset()) * 60 * 1000;
-  let oneDay = 1000 * 60 * 60 * 24;
-  let dayNumber = Math.floor(diff / oneDay) % 365;
-  return dayNumber;
-}
-
-function parseWords(words) {
-  return words.split("").map((chars) => chars.toUpperCase());
-}
-
-// Takes in a list of parsedChars e.g ["A","B","C"....]
-function generateEasyGrid(parsedChars) {
-  const result = { rows: [] };
-
-  for (let i = 0; i < 5; i++) {
-    let row = { cols: [], index: i };
-    for (let j = 0; j < 5; j++) {
-      let answer = parsedChars[i * 5 + j];
-      let value = null;
-      let state_tmp = "default";
-      // Above state types should probably be declared as a constant elsewhere
-
-      if (i === j) {
-        value = answer;
-        state_tmp = "correct";
-      }
-
-      let col = {
-        row: i,
-        col: j,
-        value: value,
-        answer: answer,
-        state: state_tmp,
-        readonly: value !== null,
-      };
-      row.cols.push(col);
-    }
-    result.rows.push(row);
-  }
-  return result;
-}
-
 function generateNormalGrid(parsedChars) {
   const result = { rows: [] };
 
@@ -74,34 +26,6 @@ function generateNormalGrid(parsedChars) {
         readonly: value !== null,
       };
       row.cols.push(col);
-    }
-    result.rows.push(row);
-  }
-  return result;
-}
-
-// Generates neighbours for the diagonal mode i.e. only considers letters of a row
-function generateRowNeighbours(gridObject) {
-  const result = { rows: [] };
-  for (let i = 0; i < 5; i++) {
-    let row = { cols: [], index: i };
-    let rowCopy = { ...gridObject }.rows[i].cols;
-
-    for (let j = 0; j < 5; j++) {
-      let charsExcludingSelf = [];
-      for (let cell of rowCopy) {
-        if (cell.col !== j && cell.answer !== cell.value) {
-          charsExcludingSelf.push(cell.answer);
-        }
-      }
-
-      let attributes = {
-        row: i,
-        col: j,
-        neighbours: charsExcludingSelf,
-      };
-
-      row.cols.push(attributes);
     }
     result.rows.push(row);
   }
@@ -159,11 +83,4 @@ function generateRowColNeighbours(gridObject) {
   return result;
 }
 
-export {
-  generateEasyGrid,
-  generateNormalGrid,
-  generateRowNeighbours,
-  generateRowColNeighbours,
-  parseWords,
-  getDayNumber,
-};
+export { generateNormalGrid, generateRowColNeighbours };
